+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Maybe:
 - add pad_gsg_short and pad_gsg_open to components
 - a function in settings exports as dict {'function': straight, 'width': 3}
     - works also for partial and composed functions
+- add `get_child_name` for Component, so that when you run `copy_child_info` the name prefix also propagates
 
 ## 3.7.3